IDSG2/IDS_Project/requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
scapy
pandas
scikit-learn
joblib
cryptography
pycryptodome
python-dotenv
plotly
matplotlib
bcrypt
argon2-cffi
reportlab
mysql-connector-python

